pyproject.toml
[project]
name = "steering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.40.0",
    "bitsandbytes>=0.49.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.8.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "steering-vectors>=0.12.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
